fix: prompted sudo password is not ignored.
If the sudo password was needed, the user was prompted, but they
password they entered never ended up getting used because of a logic
error.

The prompted password would only be used if the _sudo_password argument
was missing. If no sudo password was given, it is set to None. So
_sudo_password is present in the arguents, but does not have a value.

This fix checks if the _sudo_password argument is None and sets it using
the prompted password if so.